Move AsArg, AsVariant: object -> object/as_arg (internal-only)
