+++ gdnative-core/src/object/mod.rs
@@ -25,6 +25,7 @@ use crate::export::NativeClass;
 use crate::private::{get_api, ManuallyManagedClassPlaceholder, ReferenceCountedClassPlaceholder};
 use crate::sys;
 
+pub use as_arg::*;
 pub use instance::*;
 pub use new_ref::NewRef;
 pub use raw::RawObject;
@@ -33,6 +34,7 @@ pub mod bounds;
 pub mod memory;
 pub mod ownership;
 
+mod as_arg;
 mod instance;
 mod new_ref;
 mod raw;
@@ -304,8 +306,6 @@ unsafe impl<T: GodotObject, Own: Ownership + Send> Send for Ref<T, Own> {}
 
 /// `Ref` is `Sync` if the thread access is `Shared`.
 unsafe impl<T: GodotObject, Own: Ownership + Sync> Sync for Ref<T, Own> {}
-
-impl<T: GodotObject, Own: Ownership> private::Sealed for Ref<T, Own> {}
 
 /// `Ref` is `Copy` if the underlying object is manually-managed, and the access is not
 /// `Unique`.
@@ -994,39 +994,6 @@ impl<'a, T: GodotObject> TRef<'a, T, Shared> {
     }
 }
 
-/// Trait for safe conversion from Godot object references into API method arguments. This is
-/// a sealed trait with no public interface.
-///
-/// In order to enforce thread safety statically, the ability to be passed to the engine is only
-/// given to some reference types. Specifically, they are:
-///
-/// - All *owned* `Ref<T, Unique>` references. The `Unique` access is lost if passed into a
-///   method.
-/// - Owned and borrowed `Shared` references, including temporary ones (`TRef`).
-///
-/// It's unsound to pass `ThreadLocal` references to the engine because there is no guarantee
-/// that the reference will stay on the same thread.
-///
-/// To explicitly pass a null reference to the engine, use `Null::null` or `GodotObject::null`.
-pub trait AsArg<T>: private::Sealed {
-    #[doc(hidden)]
-    fn as_arg_ptr(&self) -> *mut sys::godot_object;
-
-    #[doc(hidden)]
-    #[inline]
-    unsafe fn to_arg_variant(&self) -> crate::core_types::Variant {
-        crate::core_types::Variant::from_object_ptr(self.as_arg_ptr())
-    }
-}
-
-/// Trait for safe conversion from Godot object references into Variant. This is
-/// a sealed trait with no public interface.
-///
-/// Used for `Variant` methods and implementations as a trait bound to improve type inference.
-pub trait AsVariant: AsArg<<Self as AsVariant>::Target> {
-    type Target;
-}
-
 /// Represents an explicit null reference in method arguments. This works around type inference
 /// issues with `Option`. You may create `Null`s with `Null::null` or `GodotObject::null`.
 pub struct Null<T>(PhantomData<T>);
@@ -1039,77 +1006,4 @@ impl<T: GodotObject> Null<T> {
     pub fn null() -> Self {
         Null(PhantomData)
     }
-}
-
-impl<'a, T> private::Sealed for Null<T> {}
-impl<'a, T: GodotObject> AsArg<T> for Null<T> {
-    #[inline]
-    fn as_arg_ptr(&self) -> *mut sys::godot_object {
-        std::ptr::null_mut()
-    }
-}
-impl<'a, T: GodotObject> AsVariant for Null<T> {
-    type Target = T;
-}
-
-impl<'a, T: GodotObject> private::Sealed for TRef<'a, T, Shared> {}
-impl<'a, T, U> AsArg<U> for TRef<'a, T, Shared>
-where
-    T: GodotObject + SubClass<U>,
-    U: GodotObject,
-{
-    #[inline]
-    fn as_arg_ptr(&self) -> *mut sys::godot_object {
-        self.as_ptr()
-    }
-}
-impl<'a, T: GodotObject> AsVariant for TRef<'a, T, Shared> {
-    type Target = T;
-}
-
-impl<T, U> AsArg<U> for Ref<T, Shared>
-where
-    T: GodotObject + SubClass<U>,
-    U: GodotObject,
-{
-    #[inline]
-    fn as_arg_ptr(&self) -> *mut sys::godot_object {
-        self.as_ptr()
-    }
-}
-impl<T: GodotObject> AsVariant for Ref<T, Shared> {
-    type Target = T;
-}
-
-impl<T, U> AsArg<U> for Ref<T, Unique>
-where
-    T: GodotObject + SubClass<U>,
-    U: GodotObject,
-{
-    #[inline]
-    fn as_arg_ptr(&self) -> *mut sys::godot_object {
-        self.as_ptr()
-    }
-}
-impl<T: GodotObject> AsVariant for Ref<T, Unique> {
-    type Target = T;
-}
-
-impl<'a, T: GodotObject> private::Sealed for &'a Ref<T, Shared> {}
-impl<'a, T, U> AsArg<U> for &'a Ref<T, Shared>
-where
-    T: GodotObject + SubClass<U>,
-    U: GodotObject,
-{
-    #[inline]
-    fn as_arg_ptr(&self) -> *mut sys::godot_object {
-        self.as_ptr()
-    }
-}
-impl<'a, T: GodotObject> AsVariant for &'a Ref<T, Shared> {
-    type Target = T;
-}
-
-mod private {
-    pub trait Sealed {}
 }
